+++ Runtime/StateGraph/States/RelayState.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using Nonatomic.VSM2.NodeGraph;
+using Nonatomic.VSM2.StateGraph.Attributes;
+
+namespace Nonatomic.VSM2.StateGraph.States
+{
+	[NodeColor(NodeColor.Blue)]
+	public class RelayState : State
+	{
+		[Transition(frameDelay:0)] public event Action OnComplete;
+		
+		public override void OnEnterState()
+		{
+			OnComplete?.Invoke();
+		}
+
+		public override void OnExitState()
+		{
+			//...
+		}
+	}
+}